Update CloudFormation templates to match lessons
The CloudFormation templates for the Advanced VPC Peering and Endpoint
lessons do not match the instructor's environment in the lesson videos.
Some of the lesson instructions cannot be completed with the provided
templates.  These updates provide the missing infrastructure and allow
the student to complete the lessons.

The following infrastructure is added to the CloudFormation templates:

* AppInstance2 EC2 instance
* SecurityGroup for instance
* IamInstanceProfile for Bastion EC2 Instance
* Rename VPC to VPC1 so consistent with VPC2
* Add missing IamInstanceProfile for AppInstance

As an example, in the VPCEndpoint video the instructor is able to query the S3 API
using the command line tools.  However, the IamInstanceProfile for
AppInstance is not present in the CloudFormation file and the S3 query
is not successful.